+++ packages/gov-ui-kit/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 -   Update minor and patch Github dependencies
 -   Update minor and patch NPM dependencies
+-   Bump `elliptic` from 6.5.7 to 6.6.0
 
 ### Fixed